+++ selfdrive/car/volkswagen/interface.py
@@ -7,8 +7,8 @@ EventName = car.CarEvent.EventName
 
 
 class CarInterface(CarInterfaceBase):
-  def __init__(self, CP, CarController, CarState):
-    super().__init__(CP, CarController, CarState)
+  def __init__(self, CP, CarController, IOCController, CarState):
+    super().__init__(CP, CarController, IOCController, CarState)
 
     self.displayMetricUnitsPrev = None
     self.buttonStatesPrev = BUTTON_STATES.copy()